Renaming of command file names
