+++ src/commands/help/command.ts
