--- src/commands/index.ts
@@ -1,4 +1,0 @@
-export { Help } from "./help/help.ts";
-export { Test } from "./test/command.ts";
-export { Watch } from "./watch/watch.ts";
-export { Command } from "./command.ts";
