+++ src/commands/watch/command.ts
